requirement.txt
absl-py==1.3.0
AmberLite==22.0
AmberUtils==21.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
astunparse==1.6.3
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
biopython==1.80
bleach==6.2.0
Bottleneck==1.3.5
brotlipy==0.7.0
bson==0.5.10
cached-property==1.5.2
cachetools==5.5.2
captum==0.2.0
certifi==2025.1.31
cffi==1.15.1
cftime==1.5.1.1
chardet==5.2.0
charset-normalizer==2.0.4
cilog==1.2.3
click==8.1.3
cloudpickle==2.2.0
comm==0.2.1
conda-pack==0.6.0
cryptography==38.0.1
cycler==0.11.0
Cython==3.0.11
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dive-into-graphs==1.0.0
docker-pycreds==0.4.0
easydict==1.13
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==1.2.0
fasteners==0.16.3
fastjsonschema==2.21.1
filelock==3.13.1
flit_core==3.6.0
fonttools==4.25.0
freetype-py==2.3.0
fsspec==2024.2.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.0
google-auth-oauthlib==0.4.6
greenlet==3.0.3
GridDataFormats==1.0.1
grpcio==1.50.0
gsd==2.6.1
h5py==3.11.0
hydra-core==1.2.0
idna==3.4
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.15.2
ipython==8.6.0
ipython_genutils==0.2.0
ipywidgets==8.1.5
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==7.4.7
jupyter_core==4.11.2
jupyter-server==1.23.4
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.2
llvmlite==0.39.1
loguru==0.7.3
lxml==5.2.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mda_xdrlib==0.2.0
MDAnalysis==2.3.0
mdtraj==1.10.0
mistune==3.1.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
MMPBSA.py==16.0
mmtf-python==1.1.3
mpmath==1.2.1
mrcfile==1.4.3
msgpack==1.0.3
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic==1.2.0
nbclient==0.6.4
nbconvert==7.16.6
nbformat==5.10.3
nest-asyncio==1.5.5
netCDF4==1.6.5
networkx==2.8.4
notebook==6.5.7
notebook_shim==0.2.4
numba==0.56.4
numexpr==2.8.4
numpy==1.23.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.2.3
openff-amber-ff-ports==0+untagged.32.g809f411.dirty
openff-interchange==0.3.25
openff-models==0.1.2
openff-toolkit==0.15.2
openff-units==0.2.2
openff-utilities==0.1.12
openforcefields==2024.9.0
OpenMM==8.0.0
openmmforcefields==0.12.0
openpyxl==3.0.10
packaging==21.3
packmol-memgen==1.2.3rc0
pandas==1.5.1
pandocfilters==1.5.0
panedr==0.8.0
ParmEd==4.2.2
parso==0.8.3
pathtools==0.1.2
pdb4amber==22.0
pdbfixer==1.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
Pint==0.23
pip==22.2.2
pkgutil_resolve_name==1.3.10
prefetch-generator==1.0.1
ProDy==2.2.0
prometheus_client==0.21.1
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3Dmol==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.25.0
pycparser==2.21
pydantic==1.10.17
pyedr==0.8.0
Pygments==2.13.0
pyMSMT==22.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyscf==1.7.6.post1
PySocks==1.7.1
python-constraint==1.4.0
python-dateutil==2.8.2
pytraj==2.0.6
pytz==2022.1
PyYAML==6.0.2
pyzmq==23.2.0
rdkit==2022.9.1
referencing==0.36.2
reportlab==4.2.2
requests==2.28.1
requests-oauthlib==1.3.1
rlPyCairo==0.2.0
rmsd==1.4
rpds-py==0.20.0
rsa==4.9
rustworkx==0.16.0
sander==22.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.3
sentry-sdk==1.10.1
setproctitle==1.3.2
setuptools==65.5.0
shap==0.41.0
shortuuid==1.0.9
sidechainnet==0.7.6
six==1.16.0
slicer==0.0.7
smirnoff99frosst==0+unknown
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.5
spyrmsd==0.8.0
SQLAlchemy==2.0.24
stack-data==0.6.0
sympy==1.11.1
tables==3.9.2
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.18.1
threadpoolctl==2.2.0
tidynamics==1.0.0
tinycss2==1.4.0
tinydb==4.8.2
torch==2.4.0+cu121
torch_cluster==1.6.3+pt24cu121
torch-geometric==2.1.0.post1
torch_scatter==2.1.2+pt24cu121
torch_sparse==0.6.18+pt24cu121
torch_spline_conv==1.2.2+pt24cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
triton==3.0.0
typed-argument-parser==1.7.2
typing_extensions==4.9.0
typing-inspect==0.8.0
urllib3==1.26.12
validators==0.34.0
wandb==0.13.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.13
xmltodict==0.14.2
zipp==3.21.0
